backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-dotenv
google-cloud-firestore
